[Settings] All options saved in spartan.xml now have names

+++ runtime/RHI/D3D12/D3D12_CommandList.cpp
@@ -48,7 +48,7 @@ using namespace Spartan::Math;
 
 namespace Spartan
 {
-    RHI_CommandList::RHI_CommandList(const RHI_Queue_Type queue_type, const uint32_t swapchain_index, void* cmd_pool, const char* name)
+    RHI_CommandList::RHI_CommandList(const RHI_Queue_Type queue_type, const uint64_t swapchain_index, void* cmd_pool, const char* name)
     {
         SP_ASSERT(cmd_pool != nullptr);
