+++ runtime/RHI/Vulkan/Vulkan_CommandList.cpp
@@ -252,7 +252,7 @@ namespace Spartan
         }
     }
 
-    RHI_CommandList::RHI_CommandList(const RHI_Queue_Type queue_type, const uint32_t swapchain_id, void* cmd_pool, const char* name) : SpObject()
+    RHI_CommandList::RHI_CommandList(const RHI_Queue_Type queue_type, const uint64_t swapchain_id, void* cmd_pool, const char* name) : SpObject()
     {
         m_queue_type  = queue_type;
         m_object_name = name;
